# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Auriga Virtual Fiducial Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /research-debt:
    get:
      operationId: listResearchDebt
      tags: [research-debt]
      summary: List research debt entries
      responses:
        "200":
          description: List of research debt entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResearchDebt"
    post:
      operationId: createResearchDebt
      tags: [research-debt]
      summary: Create a research debt entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchDebtInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchDebt"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /research-debt/{id}:
    get:
      operationId: getResearchDebt
      tags: [research-debt]
      summary: Get a research debt entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Research debt entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchDebt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateResearchDebt
      tags: [research-debt]
      summary: Update a research debt entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResearchDebtUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchDebt"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteResearchDebt
      tags: [research-debt]
      summary: Delete a research debt entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /validation/runs:
    get:
      operationId: listValidationRuns
      tags: [validation]
      summary: List synthetic validation runs
      responses:
        "200":
          description: List of validation runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValidationRun"
    post:
      operationId: createValidationRun
      tags: [validation]
      summary: Create a synthetic validation run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidationRunInput"
      responses:
        "201":
          description: Created validation run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationRun"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /validation/runs/{id}:
    get:
      operationId: getValidationRun
      tags: [validation]
      summary: Get a validation run
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationRun"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ResearchDebt:
      type: object
      required:
        - id
        - title
        - description
        - severity
        - status
        - category
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [open, in-progress, resolved, wont-fix]
        category:
          type: string
          enum: [estimation-accuracy, calibration, dataset, validation, observability, performance, other]
        relatedCalibrationProfileId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ResearchDebtInput:
      type: object
      required:
        - title
        - description
        - severity
        - category
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        severity:
          type: string
          enum: [critical, high, medium, low]
        category:
          type: string
          enum: [estimation-accuracy, calibration, dataset, validation, observability, performance, other]
        relatedCalibrationProfileId:
          type: integer
    ResearchDebtUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        status:
          type: string
          enum: [open, in-progress, resolved, wont-fix]
        category:
          type: string
          enum: [estimation-accuracy, calibration, dataset, validation, observability, performance, other]
        relatedCalibrationProfileId:
          type: integer
    ValidationRun:
      type: object
      required:
        - id
        - name
        - testType
        - calibrationProfileId
        - totalCases
        - passedCases
        - failedCases
        - passRate
        - status
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        testType:
          type: string
          enum: [lighting-variation, reflective-environment, partial-occlusion, adversarial, standard]
        calibrationProfileId:
          type: integer
        totalCases:
          type: integer
        passedCases:
          type: integer
        failedCases:
          type: integer
        passRate:
          type: number
          description: Pass rate 0-100
        status:
          type: string
          enum: [running, completed, failed]
        avgEstimatedDistanceM:
          type: ["number", "null"]
        avgConfidenceScore:
          type: ["number", "null"]
        mae:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ValidationRunInput:
      type: object
      required:
        - name
        - testType
        - calibrationProfileId
        - totalCases
      properties:
        name:
          type: string
          minLength: 1
        testType:
          type: string
          enum: [lighting-variation, reflective-environment, partial-occlusion, adversarial, standard]
        calibrationProfileId:
          type: integer
        totalCases:
          type: integer
        notes:
          type: string