# lib/api-zod/src/generated/types/validationRunInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Auriga Virtual Fiducial Engine API
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationRunInputTestType } from './validationRunInputTestType';

export interface ValidationRunInput {
  /** @minLength 1 */
  name: string;
  testType: ValidationRunInputTestType;
  calibrationProfileId: number;
  totalCases: number;
  notes?: string;
}